Change various positions and dimensions to integer from double
This includes note heights, note positions, contents height
for streamviews, positions of tracks. It makes zero sense to think of these as
having potentially fractional positions.

In addition, fractional note heights and positions lead to numerous
errors drawing MIDI stuff at the pixel level.

+++ gtk2_ardour/midi_region_view.h
@@ -177,7 +177,7 @@ public:
 	void leave_internal ();
 	void mouse_mode_changed ();
 
-	double contents_height() const { return (_height - TimeAxisViewItem::NAME_HIGHLIGHT_SIZE - 2); }
+	int contents_height() const { return (_height - TimeAxisViewItem::NAME_HIGHLIGHT_SIZE - 2); }
 
 	void connect_to_diskstream ();
 	bool pianoroll_window_deleted (GdkEventAny*);
