+++ gtk2_ardour/midi_streamview.h
@@ -69,11 +69,11 @@ public:
 	//bool can_change_layer_display() const { return false; } // revert this change for now.  Although stacked view is weirdly implemented wrt the "scroomer", it is still necessary to be able to manage layered regions.
 	void redisplay_track ();
 
-	double contents_height() const {
-		return (child_height() - TimeAxisViewItem::NAME_HIGHLIGHT_SIZE - 2);
+	int contents_height() const {
+		return child_height() - TimeAxisViewItem::NAME_HIGHLIGHT_SIZE - 2;
 	}
 
-	double y_position () const;
+	int y_position () const;
 
 	RegionView* create_region_view (std::shared_ptr<ARDOUR::Region>, bool, bool);
