+++ gtk2_ardour/midi_view.h
@@ -630,11 +630,11 @@ class MidiView : public virtual sigc::trackable, public LineMerger
 
 	virtual double height() const;
 
-	virtual double contents_height() const { return height() - 2; }
-	inline double note_height() const { return contents_height() / _midi_context.contents_note_range(); }
+	virtual int contents_height() const { return height() - 2; }
+	inline int note_height() const { return contents_height() / _midi_context.contents_note_range(); }
 
-	double note_to_y (uint8_t note) const { return _midi_context.note_to_y (note); }
-	uint8_t y_to_note (double y) const { return _midi_context.y_to_note (y); }
+	int note_to_y (uint8_t note) const { return _midi_context.note_to_y (note); }
+	uint8_t y_to_note (int y) const { return _midi_context.y_to_note (y); }
 
 	void update_patch_changes ();
 	void update_sysexes ();
